feat: 🎸 change behavior of missing config or split in rows/
Before: split was required, and config was replaced with "default" if
missing. Now: if config is missing all the rows of the dataset (all the
configs, all the splits) are returned. If split is missing, all the rows
(all the splits) are returned.

BREAKING CHANGE: 🧨 new behavior is config or split is missing in /rows