Grouped endpoints (#44)
* docs: ✏️ fix doc

* feat: 🎸 change format of /datasets response

{"dataset": name} instead of name

BREAKING CHANGE: 🧨 /datasets output changes

* refactor: 💡 Separate concepts

InfoItem might be used in a list of datasets with details, for example,
while InfoDict is the response.

* refactor: 💡 rename xxxDict to xxxContent

* refactor: 💡 remove dead code

* refactor: 💡 rename extract_rows in get_rows

It will avoid confusion, as all the other functions are get_...

* refactor: 💡 indirection for the type of every item in content

* feat: 🎸 change format of /configs response

BREAKING CHANGE: 🧨 new format

* feat: 🎸 change format of /splits output

BREAKING CHANGE: 🧨 change format

* feat: 🎸 change /rows output

BREAKING CHANGE: 🧨 change format

* feat: 🎸 change behavior of /splits if config is missing

Before, if config was missing, it was replaced by the default config
name ("default"). Now, all the configs for the dataset are fetched, then
the splits for all the configs are returned

BREAKING CHANGE: 🧨 change of behavior if config param is missing

* feat: 🎸 change behavior of missing config or split in rows/

Before: split was required, and config was replaced with "default" if
missing. Now: if config is missing all the rows of the dataset (all the
configs, all the splits) are returned. If split is missing, all the rows
(all the splits) are returned.

BREAKING CHANGE: 🧨 new behavior is config or split is missing in /rows