feat: 🎸 change the logic: decouple the /splits and /rows (#116)
* feat: 🎸 change the logic: decouple the /splits and /rows

 Now, we have jobs to get all the split names of a dataset, and other
jobs to get the rows and columns of a given split. This allows to fill
the cache progressively, and supports datasets with part of the splits
being erroneous.

BREAKING CHANGE: 🧨 the endpoints have changed, and the mongo database too

* feat: 🎸 keep splits order

* test: 💍 fix test (the splits are ordered now)

* fix: 🐛 fix doc and remove unnecessary test

* feat: 🎸 add status field to the jobs in the queue

It makes the logic easier to understand (ie. the code is easier to
maintain). Also: we give more details in the /queue and /queue-dump
endpoints, as finished jobs are now split into: error, success and
cancelled.

* feat: 🎸 give cache and cache-reports by status

instead of valid/error: empty/valid/error/stalled

BREAKING CHANGE: 🧨 two endpoints have changed

* test: 💍 fix tests