feat: 🎸 use the last revision of canonical datasets
It will correspond to the "main" branch of the canonical datasets on the
Hub, since they are mirrored. Note also that on every mirroring
operation, a webhook is sent, and the cache in the dataset viewer
backend is invalidated, in order to use the new version of the dataset.
See https://github.com/huggingface/datasets-preview-backend/issues/37
and https://github.com/huggingface/moon-landing/issues/1345 for more
details.