refactor: 💡 use datasets' get_dataset_config_info() function
We also refactor to use the info as a DatasetInfo object, instead of
serializing it to a dict (we historically did it to send it as a JSON in
a dedicated endpoint). It impacts all the column related code. Also:
disable two tests since common_voice has changed in some way that breaks
the tests. Also: upgrade datasets to get the "get_dataset_config_info"
function.