Fix detection of pending jobs (#198)
* fix: 🐛 clean DbSplitError in clean functions

* fix: 🐛 ensure the cache and queue are sync during refresh

internally, the cache and the queue have no dependency, and it's the
responsability of the workers, and of the webhook handler, to update
both accordingly. Now, an empty dataset is created in the cache when a
job is created for it, so that we can tell the user that the dataset is
being processed. In the same way, we check if the split (or even the
dataset) has a pending job if it is empty or missing in the cache, to
tell the user the split is being processed.

* style: 💄 fix style

* test: 💍 fix tests

* test: 💍 adapt test to change in dataset