fix: 🐛 quick fix to avoid mongodb errors with big rows
if a row is too big to be inserted in the cache database, we just store
the empty string for each of its cells, and mark it as erroneous. All
the cells are marked as truncated in the /rows endpoints. See
https://github.com/huggingface/datasets-preview-backend/issues/197.

This commit also contains the first migration script, with instructions
to apply and write them.

BREAKING: the cache database structure has changed. Apply the migration
script: 20220406_cache_dbrow_status_and_since.py