feat: 🎸 delete the 'rows' and 'columns' collections
instead of keeping a large collection of rows and columns, then compute
the response on every endpoint call, possibly truncating the response,
we now pre-compute the response and store it in the cache. We lose the
ability to get the original data, but we don't need it. It fixes #197.
See
https://github.com/huggingface/datasets-preview-backend/issues/197#issuecomment-1092700076.

BREAKING CHANGE: 🧨 the cache database structure has been modified. Run
20220408_cache_remove_dbrow_dbcolumn.py to migrate the database.