feat: 🎸 add NFS for /assets and dataset library /cache
Also: remove the tests for the load (cpu/ram) in the workers. Note: we
run the containers with user 1000 and group 3000, and use an
initContainer to ensure the mounted volumes from NFS use these user and
group. See https://github.com/huggingface/datasets-server/issues/241 for
a possible security improvement on this.