Nfs (#242)
* docs: ✏️ clarify doc about Release and Instance name

* fix: 🐛 use the release name to name all the objects

this way, mongo, which prefixes its name with the release, appears in
the same "group" of pods when ordered alphabetically

* fix: 🐛 remove "public," from cache-control

it is not needed (public is only meant to force public caching for
authenticated requests, which is not the case). And it prevented nginx
to correctly cache the responses based on cache-control.

* refactor: 💡 create one template directory per service

Also: use a static name for the containers (see
https://github.com/huggingface/datasets-server/commit/7abd07dc0046410adbbceda6684ea14c5d0f7b48#r73243800)

* feat: 🎸 ignore cpu and ram tests if MAX set to 0

Also: add doc to allow using HF_DATASETS_CACHE and HF_MODULES_CACHE to
set the datasets library cache directories.

* feat: 🎸 add NFS for /assets and dataset library /cache

Also: remove the tests for the load (cpu/ram) in the workers. Note: we
run the containers with user 1000 and group 3000, and use an
initContainer to ensure the mounted volumes from NFS use these user and
group. See https://github.com/huggingface/datasets-server/issues/241 for
a possible security improvement on this.