Add metrics (#258)
* fix: 🐛 remove EXPOSE since the port will depend on an env var

It does not make sense to expose a port at build time if the port will
be defined at execution time. See
https://docs.docker.com/engine/reference/run/#expose-incoming-ports

* feat: 🎸 add an endpoint for the prometheus metrics

* feat: 🎸 add a metrics endpoint

BREAKING CHANGE: 🧨 environment variables have changed

* feat: 🎸 use the new env vars format in Kube

* fix: 🐛 move the metrics into the API as /metrics

It makes no sense to have it in a parallel application.

* feat: 🎸 rename api to app

* feat: 🎸 create the /metrics endpoint and remove /queue

The queue information is now available as Prometheus metrics in /metrics

* feat: 🎸 add cache metrics, and remove /cache endpoint

* test: 💍 fix tests

* feat: 🎸 rename /metrics in /prometheus

It's clearer, and it will allow to provide info about the dataset
metrics (https://huggingface.co/metrics) if needed. Also: fix tests

* docs: ✏️ update the readme

* test: 💍 launch mongo before the tests

* test: 💍 fix tests

(and style)