feat: 🎸 use "stale" instead of "stalled" in the cache db
Fixes https://github.com/huggingface/datasets-server/issues/368.

"stalled" is still allowed in the database and is considered as "stale",
but it will not be written anymore.
The value "stalled" can be replaced with "stale" in the database, but
it's not mandatory.

BREAKING CHANGE: 🧨 method names use "stale" instead of "stalled"