feat: 🎸 change the response of /cache-reports
only report for the new cache entries: /splits-next and /first-rows,
since /splits and /rows will be deprecated soon. Also: reduce the size
of the response to avoid timeouting. The client can fetch the complete
response by querying the corresponding endpoint if needed (to get the
cause message or the cause traceback for example)