Improve technical routes response (#454)
* refactor: 💡 group the routes and add comments

* feat: 🎸 deprecate /hf_datasets

It serves no purpose since we can get more with
https://huggingface.co/api/datasets

* perf: ⚡️ reduce the response size

* feat: 🎸 remove /hf-datasets-count-by-cache-status

we will instead expose the cache contents, and let the API users do this
kind of analysis by theirselves

* feat: 🎸 deprecate /refresh-split endpoint

/webhook is the way to do, and we refresh a dataset, not a specific
split (it's too internal to expose)

* feat: 🎸 deprecate /queue-dump

it is too heavy (the server times out) because it returns the whole list
of the past jobs. It was only used externally to do stats about the
duration of the jobs, which is not that important, and can be added to
prometheus better if needed.

* feat: 🎸 rename /queue-dump-waiting-started in /pending-jobs

for clarity

* feat: 🎸 change format of /pending-jobs and add two queues

* refactor: 💡 try to describe the routes better

* feat: 🎸 add get_splits_response_reports (same for first-rows)

* feat: 🎸 change the response of /cache-reports

only report for the new cache entries: /splits-next and /first-rows,
since /splits and /rows will be deprecated soon. Also: reduce the size
of the response to avoid timeouting. The client can fetch the complete
response by querying the corresponding endpoint if needed (to get the
cause message or the cause traceback for example)