feat: 🎸 add a script to refresh the canonical datasets (#463)
* feat: 🎸 add a script to refresh the canonical datasets

it's useful to relaunch the jobs on the canonical datasets after
upgrading the datasets library in service/worker. Indeed, the canonical
datasets are versioned in the same repo as the datasets library, and are
thus expected to work with that version. Also: every new release of
datasets triggers an update (synchro) of every canonical dataset on the
hub, which triggers the datasets-server webhook and adds a refresh job
for every canonical dataset. As they are still run with the outdated
version of the datasets library, it's good to refresh them once the
library has been upgraded

* ci: 🎡 fix yaml filenames

(it's important because there were a mismatch between the name and the
rule to run them)