Add auth to api endpoints (#495)
* feat: 🎸 add auth_check to check if a dataaset can be accessed

* feat: 🎸 add external authentication to two endpoints

/first-rows and /splits-next

* fix: 🐛 the url can be None

* fix: 🐛 return 404 when external auth fails with 403

When it's 401, we return 401 because authenticating might fix the issue.

* docs: ✏️ add authentication mechanisms and update paths

* docs: ✏️ document /healthcheck, and add 500 Internal error

If starlette crashes, it returns a 500 error in text/plain. Note that if
starlette crashes without being able to send a response, uvicorn will
return a 500 error too, with the exact same content ("Internal Server
Error").

* feat: 🎸 protect gated datasets for /is-valid

* feat: 🎸 update API docker image, add e2e tests, fix openapi

also: order the e2e test files alphabetically to be able to use the
result of past tests (not sure if it's a good idea). Also: pass
EXTERNAL_AUTH_URL to the API service in docker compose and helm.

* test: 💍 re-enable tests

* ci: 🎡 ignore safety warnings

ujson is a dependency of lm-dataformat, which last release was one year
ago.

* ci: 🎡 fix a typo

* docs: ✏️ fix healthcheck response

* test: 💍 add a message if assert fails, to help debug

The issue here is that e2e works locally, and not in the CI. This will
help debug.

* test: 💍 add more assert messages

* test: 💍 fix parenthesis

* ci: 🎡 add the URL to the hf.co auth check endpoint

* feat: 🎸 update the docker images

(and fix the file, btw)

* feat: 🎸 remove the reverse-proxy cache

All the requests go to the api/admin service directly. This allows to do
an auth check for every request.