fix: 🐛 respect HF_ENDPOINT when testing if dataset exists
instead of relying on huggingface_hub and having to HF_ENDPOINT to it,
we just rely on datasets, which raises an exception in that case (we
have to rely on the error message, which is not very good). Also:
disable tests for the moment, because the datasets have been deleted.