feat: 🎸 add (/splits) jobs if a /first-rows resp is cache miss
If a request is made to /first-rows and there is no entry in the
cache, we check if it should.

Various cases have to be managed, which makes the logic a bit (too)
convoluted.

We return a FirstRowsNotReady error if:
- the /first-rows response for the split is already in process
- the /splits response for the dataset is still in process

We return a FirstRowsNotReady error, AND we add a /splits job because
something has failed before, if:
- the /splits response for the dataset exists, and the split is part
  of the splits of the dataset
- the /splits response for the dataset does not exist, is not in
  process, but it should because the dataset is supported.

Note that I didn't add a unit test for every of these cases