feat: 🎸 make the queue agnostic to the types of jobs
Before we had two collections: for splits and for first-rows jobs. Now
only one collection name "jobs", with a field "type". Note that the job
arguments are still restricted to dataset (required) and optionally
config and split.

BREAKING CHANGE: 🧨 two collections are removed and a new one is created. The function names
have changed too.