refactor: 💡 setup everything in the configs
the connection to the databases, the logging configuration and the
creation of the assets directory are done automatically just after
getting the configuration, removing the need to do it when starting an
app. Also: simplify the use of logging, by just calling the
logging.xxx() functions directly, instead of using a custom logger.