feat: 🎸 limit the started datasets per namespace
and also: select the next dataset among the namespaces with the least
number of started jobs, in order to avoid having all the workers
dedicated to the same user, if others are waiting. The namespace is the
user or the org, or the dataset name itself for canonical datasets.

Deployment: the collection is now called jobs_blue (see blue/gree
deployment). Once deployed, look at the old "jobs" collection, relaunch
the jobs for the waiting or started datasets, then delete the "jobs"
collection.