feat: 🎸 replace Queue.add_job with Queue.upsert_job
upsert_job ensures there is only one waiting job for the same set of
parameters. On every call to upsert_job, all the previous waiting jobs
for the same set of parameters are cancelled, and a new one is created,
with a new "created_at" date, which means it will be put at the end of
the queue. It will help to fight against datasets that are updated very
often (eg every minute): they will be treated only when there is
available workers.