refactor: 💡 set libcommon as an "editable" dependency
All the jobs, services and workers share the same version of
libs/libcommon, which is the current version. Before, we had to update
the libcommon version, run poetry build, update the version in the other
projects' pyproject.toml and poetry update it. The workflow when
updating libcommon will be a lot simpler now. Also: vscode will be able
to follow the reference to the source of the libcommon code, instead of
showing the packaged dependency in .venv/. Caveat: this also means that
now, modifying something in libcommon will affect all the jobs, services
and workers, so we need to have to be more careful. This also means that
more CI jobs will be run on every libcommon update, and that when
deploying on prod, all the images will be updated and redeployed.