feat: 🎸 add a /backfill admin endpoint (#708)
* feat: 🎸 add a /backfill admin endpoint

It only updates all the datasets of the Hub, with a low priority. Note
that most of the jobs will be skipped, because the response will already
be in the cache. Also thiscommit fixes the creation of children jobs:
the priority is preserved (ie low priority jobs created children low
priority jobs)

* style: 💄 fix style

* feat: 🎸 update docker image