fix: 🐛 release the resources on shutdown
they were released just after the app has been returned. But we want the
resources to be released when the app is shut down instead.

Note that as we need the resources (mainly the log, and the assets
directory for the prometheus endpoint), we need to allocate them at the
start of the function, instead of using the "on_startup" parameter of
the Starlette constructor.