fix: 🐛 allocate assets storage for all the workers
It fixes the issue raised by @lhoestq at
https://github.com/huggingface/datasets-server/pull/784#discussion_r1100498336.

Note that it's a bit radical: all the workers now require write access
to the assets storage directory, even if they will not use it.
I think it's OK since I will be working very soon on
https://github.com/huggingface/datasets-server/issues/737, which will
merge all the workers in only one worker.