feat: 🎸 reduce the requests (and limits) in k8s resources (#886)
* feat: 🎸 reduce the requests (and limits) in k8s resources

It's required for the api/admin services which were absolutely
over-provisioned. And for the workers: we reduce a lot for now, since
most of the datasets don't require that quantity of RAM, and we now
manage the failed jobs due to OOm. We will work on a strategy to
relaunch these jobs on "big" workers (to be done)

* fix: 🐛 request 1 cpu by default

only keeping 200m for really small workers