fix: 🐛 another fix on parsing the parquet file names
Because the previous fix (to support builder names with dashes into
them) was breaking the detection of the shard number. Note that we
cannot support split names that contain dashes!!! I think it's a
limitation, maybe we should store each split in its own directory
instead of trying to parse.