refactor: 💡 use networkx to create and check the graph
It removes grpah logic from the code, and will make it easier to later
allow multiple parents. Note that now the children and the ancestors
have no particular order (they could be sets instead of lists by the
way). The APIs of the classes have not changed, to make the upgrade
smoother.