feat: 🎸 get the dataset state, and backfill the missing parts (#1045)
* feat: 🎸 get the dataset state, and backfill the missing parts

* feat: 🎸 add backfill to DatasetState

It ignores the relationships config->dataset, split->config and
split->dataset.

* refactor: 💡 simplify the code

* fix: 🐛 avoid recursion

and precompute the state of all the steps in DatasetState

* refactor: 💡 simplify the code

* refactor: 💡 simplify the code

* refactor: 💡 rename StepState to ArtifactState

An artifact will now be a step for specific set of inputs (a dataset, a
config, a split). A step have multiple artifacts, depending on its
InputType. An artifact is linked to jobs, and to a cache entry

* refactor: 💡 change terms

* feat: 🎸 check the job runner version, plan refresh is outdated

* feat: 🎸 refresh if git revision is diff. from previous artifact

Note that it can lead to a bug: we need to pass the git revision to the
job, or we could have loops.

* test: 💍 add a test for artifacts that must be retried

due to error code

* refactor: 💡 move formatting methods to state.py

* feat: 🎸 add /dataset-state endpoint

It might be confusing with /dataset-status... I think the second one
will disappear once we consolidate the DatasetState code (in an
orchestrator?)

* feat: 🎸 add POST /dataset-backfill?dataset=

to backfill a dataset, using the plan elaborated from the dataset state

* fix: 🐛 pass the git revision to DatasetState

getting the current git commit of main and passing it to DatasetState
helps identifying which cache entries are outdated.

* feat: 🎸 return list of canceled jobs

It helps to factorize code in upsert_job()

* refactor: 💡 refactor tests to remove duplicated code

* fix: 🐛 fix CI