Update backfill job, and setup a cronjob in prod (#1077)
* feat: 🎸 create children jobs even in case of error

fixes #949, along with #1066

* feat: 🎸 don't delete old cache entries at the end of a job

Because a job has not the sufficient information to know which cache
entries whould be deleted or not. Let's use backfill for that (in a next
PR. It's not an issue to have dangling cache entries meanwhile)

* style: 💄 remove unnecessary imports

and delete unused code

* style: 💄 remove mypy warnings

* docs: ✏️ fix readme

* feat: 🎸 use the new DatasetState.backfill() method to backfill

Also: remove the "upgrade" action, since it's handled by backfill. Also:
remove test for backfill action, because it's not clear how it would
work (maybe -> e2e test)

* feat: 🎸 add backfill cronjob

* Update jobs/cache_maintenance/README.md

Co-authored-by: Andrea Francis Soria Jimenez <andrea@huggingface.co>

* fix: 🐛 remove duplicated line

---------

Co-authored-by: Andrea Francis Soria Jimenez <andrea@huggingface.co>