feat: 🎸 ignore result of job runner if job has been canceled (#1188)
* test: 💍 fix test

the length has changed

* refactor: 💡 move all JobRunnerError to libcommon

Two ideas behind the refactoring:
1. having libcommon aware of the list of error codes, to be able to
   relate with list of error codes to retry, in a next PR
2. centralize all the error codes, since they share the same namespace.
   It will help do support.

* fix: 🐛 fix imports

* test: 💍 fix the tests

* feat: 🎸 replace PreviousStepError with CachedArtifactError

the logic is now a lot simpler

* docs: ✏️ harmonize the docstrings

* feat: 🎸 put revision as a job param

* feat: 🎸 remove deprecated index, and add new indexes for Job

* feat: 🎸 add field revision to jobs

set the value to "main", it should do the trick for the existing jobs,
while the next ones will use the commit hashes.

* feat: 🎸 allow only one pending job per "dataset/config/split"

ie. creating a new job with a different revision will replace the
previous job

* style: 💄 fix rebase

* test: 💍 rename test

* refactor: 💡 move all JobRunnerError to libcommon

Two ideas behind the refactoring:
1. having libcommon aware of the list of error codes, to be able to
   relate with list of error codes to retry, in a next PR
2. centralize all the error codes, since they share the same namespace.
   It will help do support.

* feat: 🎸 replace PreviousStepError with CachedArtifactError

the logic is now a lot simpler

* feat: 🎸 ignore result of job runner if job has been canceled

also: refactor to remove two queue methods (kill_zombies,
kill_long_jobs): job_manager is now in charge of finishing the jobs, and
updating the cache (if needed).