Revert "fix: 🐛 finish the job before backfilling, to get the status (#1252)"
This reverts commit 1cbd9ede2ea7de7f93662c0e802cb77d378eac3c.

The backfill() step still lasts too long in the workers, for datasets
with a lot of configs/splits, leading to concurrency issues. As we don't
have prometheus metrics for the workers, we cannot benchmark on prod
data.

Reverting, and I will find another solution to finish the jobs with the
right status.