feat: 🎸 disable the limit per namespace
(only increased to 1000, which is a lot more than the max number of
started jobs). I think that we don't need this restriction anymore,
because the jobs are limited in time, and the datasets with less started
job are already prioritized. I think that the code could be removed at
one point (let's first see how it goes), which would simplify the
queries to mongo