feat: 🎸 start the newest job for unicity_id + cancel the others
it's done under the lock, so, we ensure we always have at most 1 started
job for the same unicity_id, and that all the other waiting jobs for the
same unicity_id are canceled. Also note that meanwhile, a new job can be
created (by the /webhook endpoint for example).