Ensure only one job is started for the same unicity_id (#1420)
* feat: 🎸 add a global lock on start_job and upsert_job

to avoid duplicate jobs, and multiple job runners getting the same job
at the same time.

* refactor: 💡 rename job_id to owner in Lock

includes a DB migration

* feat: 🎸 lock the started job

to start a job, the lock for the unicity_id must be acquired. If so, the
lock is owned by the job_id (it's unique). Else, the worker loop will
sleep and retry to get a job, as when the queue is empty.

* feat: 🎸 remove constraint on waiting jobs

now: we can have as many waiting jobs as we want. Queue.upsert_job
disappears, and Queue._add_job becomes public Queue.add_job. Note that
the current behavior is that every waiting job will be processed. It's
inefficient and will be fixed in another commit.

* refactor: 💡 remove dead code

* fix: 🐛 fix syntax

* fix: 🐛 ensure the job is still waiting before starting it

* feat: 🎸 start the newest job for unicity_id + cancel the others

it's done under the lock, so, we ensure we always have at most 1 started
job for the same unicity_id, and that all the other waiting jobs for the
same unicity_id are canceled. Also note that meanwhile, a new job can be
created (by the /webhook endpoint for example).

* refactor: 💡 better name